+++ CHANGELOG.md
@@ -165,3 +165,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Replace email context variables in email subject
 - Default and Default var in html tags
 - Fetch human readable bridgehead for emails
+- Email context project bridgeheads variable